backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
flower==2.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
openai-whisper==20231117
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2
ffmpeg-python==0.2.0
requests==2.31.0
httpx==0.26.0
minio==7.2.3
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

